# api/openapi3.yaml
openapi: 3.0.0
servers:
- url: http://localhost:5000/
info:
  title: Hackspace Status API
  description: This API allows to control your local hackspace status
  version: "1.0"
paths:
...
  /api/v1/sensors/temperature:
    put:
      tags:
        - Sensors
      operationId: app.status.setTemperature
      summary: Set sensors of the hackspace
      responses:
        200:
          description: OK (successfully authenticated)
        401:
          description: Unauthorized (unsuccessfully authenticated)
      requestBody:
        content:
          application/json:
            schema:
              x-body-name: received
              $ref: '#/components/schemas/sensorsTemperature'
      security:
        - api_key: []
components:
  schemas:
...
    temperature:
      type: object
      required:
        - value
        - unit
        - location
      properties:
        value:
          type: number
          description: The sensor value
          example: 16.5
        unit:
          type: string
          description: The unit of the sensor value.
          enum: ['°C', '°F', 'K', '°De', '°N', '°R', '°Ré', '°Rø']
          example: °C
        location:
          type: string
          description: The location of your sensor such as Outside, Inside, Ceiling, Roof or Room 1.
          example: Outside
        name:
          type: string
          description: This field is an additional field to give your sensor a name. This can be useful if you have multiple sensors in the same location.
        description:
          type: string
          description: An extra field that you can use to attach some additional information to this sensor instance.
...
    sensorsTemperature:
      type: object
      required:
        - sensors
      properties:
        sensors:
          type: object
          required:
            - temperature
          properties:
            temperature:
              type: array
              description: Temperature sensor.
              items:
                $ref: '#/components/schemas/temperature'
  securitySchemes:
    api_key:
      type: apiKey
      name: X-Hackspace-API-Key
      in: header
      x-apikeyInfoFunc: app.apikey.auth